Add the native visual census and L03 pet and power respec lanes
The working-save visual census opens the actual level-up screen once per
published class and ancestry (at zero investment and again when a Full
step is next), shows the favored-class reward selector through the
screen's own collection switch with the ShowAll filter, and checks every
rendered row: its exact icon, never the acronym placeholder, no empty
selector, and every leaf the native backend offers rendered. It also
opens the actual creator's Heritage phase for each Mostly Human parent
race. Coverage lists any published leaf never rendered.

L03: an Ifrit Sorcerer 6 whose native picks raised its Elemental Ray
(Fire) is respecced to Sorcerer 1; in the working save an Oread Ranger 4
with projected companion armor is respecced, its old companion destroyed
without an orphaned projection, and the counters earned again project
once onto the new companion. The respec helpers level any class through
its own reward selection, prefer the choices a subject needs and fill
spells known.

Lanes: an ended area gets the native controller's final tick before its
effects are checked, and reloaded area numbers are compared as the
doubles the reloaded JSON reads back.

+++ src/KingmakerGunslinger/RuntimeTesting/RuntimeTestRunner.FavoredClassPersistenceFamilies.cs
@@ -276,11 +276,13 @@ namespace KingmakerGunslinger.RuntimeTesting
                 var cylinder = area.View == null ? null : area.View.Shape as ScriptZoneCylinder;
                 GameObject ring = area.View == null ? null : typeof(Kingmaker.View.MapObjects.AreaEffectView)
                     .GetField("m_SpawnedFx", BindingFlags.Instance | BindingFlags.NonPublic).GetValue(area.View) as GameObject;
+                // Doubles rounded to three places: the verify process compares
+                // them with the values its reloaded JSON reads back.
                 areas.Add(new JObject
                 {
                     ["area"] = area.Blueprint.name,
-                    ["radius"] = cylinder == null ? -1f : (float)Math.Round(cylinder.Radius, 3),
-                    ["ringFactor"] = ring == null ? 1f : (float)Math.Round(FavoredClassPerformanceRing.FactorOf(ring), 3),
+                    ["radius"] = cylinder == null ? -1d : Math.Round((double)cylinder.Radius, 3),
+                    ["ringFactor"] = ring == null ? 1d : Math.Round((double)FavoredClassPerformanceRing.FactorOf(ring), 3),
                 });
             }
             result["areas"] = areas;
